Refactor and move code to New Searchbar component
