--- src/components/Navbar.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Navbar = () => {
-  return <div>Navbar</div>;
-};
-
-export default Navbar;
